# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Sijunjung Go application with authentication support",
        "title": "Sijunjung Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/account": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently delete the current user account and all associated data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Delete user account",
                "responses": {
                    "200": {
                        "description": "Account deleted successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Failed to delete account",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/facebook": {
            "post": {
                "description": "Authenticate or register user with Facebook access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authenticate with Facebook",
                "parameters": [
                    {
                        "description": "Facebook auth request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FacebookAuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.FacebookAuthData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Facebook token",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/google": {
            "get": {
                "description": "Redirects user to Google OAuth consent page for login/registration",
                "tags": [
                    "auth"
                ],
                "summary": "Redirect to Google OAuth",
                "parameters": [
                    {
                        "type": "string",
                        "description": "State parameter for CSRF protection",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to Google OAuth"
                    }
                }
            }
        },
        "/api/auth/google/callback": {
            "get": {
                "description": "Handle callback from Google OAuth and return JWT token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Google OAuth callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization code from Google",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "State parameter for CSRF protection",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.GoogleAuthData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid authorization code",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "description": "Authenticate user with email and password to get bearer token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.TokenData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoke the current bearer token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the current authenticated user information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "Current user info",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.UserData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "description": "Register a new user with full name, email, and password. Sends OTP to email for verification.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created and OTP sent",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request or email already exists",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/resend-otp": {
            "post": {
                "description": "Resend the OTP code to email (1 minute cooldown)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Resend OTP code",
                "parameters": [
                    {
                        "description": "Resend OTP request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ResendOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP resent successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "User not found or cooldown active",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Please wait before requesting new code",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/reset-password": {
            "post": {
                "description": "Reset password by sending a new random password to user's email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Reset password",
                "parameters": [
                    {
                        "description": "Reset password request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New password sent to email",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "User not found or invalid request",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/verify-otp": {
            "post": {
                "description": "Verify the OTP code sent to email and get bearer token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify OTP code",
                "parameters": [
                    {
                        "description": "Verify OTP request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.VerifyOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verification successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/http.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.TokenData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid or expired code",
                        "schema": {
                            "$ref": "#/definitions/http.APIErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.APIErrorResponse": {
            "description": "Standard API error response",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "invalid request"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "http.APIResponse": {
            "description": "Standard API response",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "http.FacebookAuthData": {
            "description": "Facebook auth data",
            "type": "object",
            "properties": {
                "is_new_user": {
                    "type": "boolean",
                    "example": true
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "http.GoogleAuthData": {
            "description": "Google auth data",
            "type": "object",
            "properties": {
                "is_new_user": {
                    "type": "boolean",
                    "example": true
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
...
        "http.TokenData": {
            "description": "Token data",
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "http.UserData": {
            "description": "User data",
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                }
            }
        },
        "model.FacebookAuthRequest": {
            "description": "Request body for Facebook authentication",
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "EAABsbCS1..."
                }
            }
        },
        "model.LoginRequest": {
            "description": "Request body for user login",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "sijunjunggo@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
...
        "model.ResendOTPRequest": {
            "description": "Request body for resending OTP",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "sijunjunggo@gmail.com"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}